[ZH] Move intrin_compat.h to Dependencies #479
