+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWDebug/wwdebug.h
@@ -43,7 +43,7 @@
 
 // TheSuperHackers @todo Recover WWDEBUG?
 #ifdef WWDEBUG
-#include <Lib/intrin_compat.h>
+#include <Utility/intrin_compat.h>
 #endif
 
 // The macro MESSAGE allows user to put:
